Autofill design/ship review feedback (#2926)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI
changes** section
-->

Task/Issue URL:
https://app.asana.com/0/488551667048375/1204135679490431/f

### Description
Address comments raised in design and ship reviews. This is the base of
a stacked PR where each individual PR on the stack was addressing one or
more items to address as part of the design / ship review.

Translations will follow in a separate PR.

### Steps to test this PR

This is the base of a stacked PR; all PRs on the stack have been
individually reviewed and approved. So this is really `QA-Optional` at
this stage.

Smoke testing autofill would be the most helpful here.

---------

Co-authored-by: Marcos Holgado <marcosh@duckduckgo.com>