Sync: DataObserver should not be triggered when sync in progress (#3295)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
    3. Make sure these changes are tested in API 23 and API 26
If your PR does not involve UI changes, you can remove the **UI
changes** section
-->

Task/Issue URL:
https://app.asana.com/0/414730916066338/1204942882074644/f

### Description
When Sync is in progress, we shouldn’t listen to underlying changes.
This is because when a sync is in progress, it could be modifying the
SavedSites. The Observer is then triggered since changes are applied to
the db.