Fix wrong placeholder icon radius (#3604)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
If your PR does not involve UI changes, you can remove the **UI
changes** section

At a minimum, make sure your changes are tested in API 23 and one of the
more recent API levels available.
-->

Task/Issue URL:
https://app.asana.com/0/488551667048375/1205173497854945/f

### Description
* Do not round corners when creating placeholder icon, except for the
widget
* Generally, we rely on the surface displaying the placeholder to do the
corner rounding, however, this is not possible on the widget due to some
platform limitations in API < 31. That's why the widget needs to apply a
different treatment. Up until now, all placeholders were provided with a
radius optimized for the widget, which caused them to have a different
shape than the actual favicons
* Fix wrong radius in widget background and icons. Shapes were not
aligned, and also not aligned with de designs in Figma

### Steps to test this PR

_Placeholder favicon for favorites, fireproofed and tab switcher_
- [x] Save a site without a favicon as favorite (or force fallback to
placeholder, instructions below). Also fireproof it and leave it open in
a tab
- [x] Save a site with a favicon. Also fireproof it and leave it open in
a tab
- [x] Check that both placeholder and favicon have the same shape in
bookmarks, favorites (clear address bar), fireproofed sites, tab
switcher, and widgets

_Placeholder favicon for location permissions_ 
- [x] Force Google maps to fallback to placeholder
- [x] Open Google maps in the browser and click on "Your location"
- [x] Check that icon with and without fallback has the same shape

#### How to force fallback to placeholder 
Example for wikipedia

in `FavoritesWidgetService#onDataSetChanged()` replace `url = it.url`
with ` url = if (it.url.contains("wikipedia")) "" else it.url`

in `FaviconImageView.ImageView.loadFavicon( bitmap: Bitmap?, domain:
String, placeholder: String? = null)` replace `.load(bitmap)` with
`.load(if (domain.contains("wikipedia")) null else bitmap)`



### UI changes
## Bookmarks

![bookmarks](https://github.com/duckduckgo/Android/assets/6297834/c1aebe25-777e-457a-ace9-60ac6ebc5940)
## Favorites

![favs](https://github.com/duckduckgo/Android/assets/6297834/05233801-8723-482e-b8ac-d247c4870f3d)
## Fireproofed sites

![fireproof](https://github.com/duckduckgo/Android/assets/6297834/0d5dc5ac-8b64-47d9-9727-fc033e488c96)
## Location permissions dialog

![locationpermissions](https://github.com/duckduckgo/Android/assets/6297834/7a2b2389-8f75-405d-8f6e-bb6c0c234e78)

![locationpermissionsplaceholder](https://github.com/duckduckgo/Android/assets/6297834/a56e4c20-29eb-4dcc-a0ba-96c62c8dc28d)
## Tab switcher

![tabswitcher](https://github.com/duckduckgo/Android/assets/6297834/fe3ac692-8b9c-4d0b-b692-92511ac5a30e)
## Widgets (background color manually set to yellow to better illustrate
the fact that shapes were not aligned)

![widgets](https://github.com/duckduckgo/Android/assets/6297834/930a5190-0629-4fad-aac1-2a0e18f1209d)