Fix viewport (#3716)
Task/Issue URL:
https://app.asana.com/0/1163321984198618/1205404061469342/f

### Description
This PR is a first attempt to fix our viewport issues

### Steps to test this PR

- [x] Install the app, go to serp and when you see the onboarding dialog
select to hide tips forever
- [x] Go to twitter.com
- [x] You should see the whole cookies popup at the bottom of the screen
and all of it should be visible.
- [x] Scrolling up and down should hide and show the URL as expected and
the cookies popup should (more or less) stay in place (it will jump a
bit).
- [x] On the same tab go to `reddit.com`, a container should be shown at
the bottom of the screen and all of it should be visible. Scroll up and
down and it should still be there.
- [x] Open a new tab and go to `reddit.com`, a container should be shown
at the bottom of the screen and all of it should be visible. Scroll up
and down and it should still be there.
- [x] On the same tab go to `twitter.com` and you should see the whole
cookies popup at the bottom of the screen and all of it should be
visible.