Never save passwords for this site: api and storage (#3998)
<!--
Note: This checklist is a reminder of our shared engineering expectations. 
The items in Bold are required
If your PR involves UI changes:
    1. Upload screenshots or screencasts that illustrate the changes before / after
    2. Add them under the UI changes section (feel free to add more columns if needed)
If your PR does not involve UI changes, you can remove the **UI changes** section

At a minimum, make sure your changes are tested in API 23 and one of the more recent API levels available.
-->

Task/Issue URL: https://app.asana.com/0/0/1206017917054330/f 

### Description
This stack of PRs will offer users a way of saying to _never ask to save a password for this site again_. This particularly PR focusses on:
- The underlying DB and the data APIs for adding sites to the excluded list, resetting the list and getting the count from the list.
- It also includes additional dev settings related to the "never saved" sites

### Steps to test this PR
Most of the testing will happen in the PR above, where the UI-based live. For this PR, just need to verify the DB migration happens ok.

#### Database upgrade
There is a database schema change, so need to check the migration is ok.

- [ ] Install from `develop`
- [ ] Add a password (overflow->Passwords, then add manually using ➕ button)
- [ ] Install from this branch. Open Passwords again, and verify the saved password is still there and correct

#### Fresh install
- [ ] Uninstall and re-install from this branch to get a fresh install. Just verify this launches ok to check the DB creation.

#### [Optional] Test dev settings
- [ ] Visit autofill dev settings (`Settings->Autofill dev settings`) and scroll to the bottom to `Never Saved Sites` section
- [ ] Tap on `Add sample site (fill.dev)`. Verify the number of excluded sites changes from 0 to `1`
- [ ] Tap it multiple times; verify it remains `1` (the same site shouldn't be added multiple times)
- [ ] Tap the `Reset Excluded Sites` button and verify it becomes `0`


---
- To see the specific tasks where the Asana app for GitHub is being used, see below:
  - https://app.asana.com/0/0/1206071669665237