Split apart AutofillStore, moving what's not needed as public to another interface/class (#4093)
<!--
Note: This checklist is a reminder of our shared engineering
expectations.
The items in Bold are required
If your PR involves UI changes:
1. Upload screenshots or screencasts that illustrate the changes before
/ after
2. Add them under the UI changes section (feel free to add more columns
if needed)
If your PR does not involve UI changes, you can remove the **UI
changes** section

At a minimum, make sure your changes are tested in API 23 and one of the
more recent API levels available.
-->

Task/Issue URL: https://app.asana.com/0/0/1206383428535278/f 

### Description

`AutofillStore` was designed as a single place to refer to perform CRUD
operations pertaining to autofill (e.g., saving a new credentials,
deleting a credential etc...).

Over time it's become a large API and, given other work to move more of
autofill into `autofill-impl`, most of the functions are now used only
by `autofill-impl` and not wider in the codebase, so most of the
functions are in the public API and don't have to be. This PR cleans
that up.

Some functions are called from elsewhere (namely, app module) so can't
fully move `AutofillStore` as a whole to autofill-impl. Instead, need to
break the API into two.


### Steps to test this PR
 Mostly just smoke testing needed here.

- [x] install from `develop` and save a few logins
- [x] install this branch and make sure they're still available and
autofill still working ok (e.g., test it out
https://fill.dev/form/login-simple)