load privacy plugins data into memory just for main process (#4189)
Task/Issue URL: https://app.asana.com/0/488551667048375/1206608309382832/f

### Description
Many privacy plugins data is loaded into memory for both `:app` and `:vpn` processes.
Even when the plugins are only used from the app process, because privacy config is only downloaded from the `main` process, the loading mistakenly happens in the `vpn` process as well because
all plugins have a `loadToMemory` method that is executed in the `init` block (ie. when the instance is created) and the `PluginPoint<MainProcessLifecycleObserver>` is a dependency to the `Application` instance,
which is created for both `vpn` and `app` processes.

However loading into data for the `vpn` process should not be required/done as we don't use that data from the `vpn` process.

This task is a stop gap to avoid that in-memory data loading to happen. 

In sub-sequent PRs we'll want to have a better nad more standardize mechanism to perform in-memory Cache backed by disk storage.


### Steps to test this PR
Code review and smoke test the app
Make sure privacy config downloads and sets correctly.