Fix OOM crash on Timber.d (#4218)
Task/Issue URL: https://app.asana.com/0/488551667048375/1206718079780225/f

### Description
Fix a OOM crash that happens in production when using Timber.d() and parameters are larged. In this case a url that embeds a file to download

The fix is replacing Timber by logcat, which is something we wanted to do anyways little by little.

### Steps to test this PR
NA, just code review