Update fastlane to upload APK straight to production to reduce wasted time on double reviews (#4566)
Task/Issue URL:
https://app.asana.com/0/488551667048375/1207360448878908/f

### Description
We experience lengthy Play Store reviews and pay the price twice for
each release:
- one review is triggered when we upload our APK to the `staging` track
- a second review is triggered when we promote from `staging` to begin
the `production` rollout

Each of the reviews can amount to many hours each. This PR changes it so
that we skip `staging` and go straight to production at 0.01% rollout so
that the review clock starts immediately and we only have to wait for a
single app review instead of needlessly waiting on two for the same
release.

### Steps to test this PR
- [x] Sense-check the change. Can't really test until we next do a
release.