Tab manager crash fix (#4878)
Task/Issue URL:
https://app.asana.com/0/1207418217763355/1208000508832396/f

### Description

This is the 2nd attempt at fixing the crash. I wasn’t able to reproduce
it, but the theory is that the adapter list gets modified while it’s
being diffed, which results in an inconsistency and
`IndexOutOfBoundsException` being thrown.

The fix for this is to create a local copy of the diffed lists, which
should resolve race condition issue. I also added index range checks to
be extra careful even though they *should* be redundant.

### Steps to test this PR

I wasn’t able to reproduce the crash. Smoke-testing the drag & drop
feature should be enough.