fix: clicking bookmark after launched not opening bookmark site
It looks like startActivity was removed when Bookmarks were moved to the saved-sites module. While bookmarks work fine if the user is has opened the app (we set a result instead) when launching from a shortcut we don't have a BrowserActivity stacked (I'm not sure if this was intended or not) so we need to launch the BrowserActivity here.

This does not solve the up button or back button issue closing the app instead of going to the home screen.