fix: pressing back or up button closing the app
We can stack BrowserActivity before the BookmarksActivity so that when users press back or the up button they end up on the browser screen rather than the app closing.

This likely worked before the previously mentioned module move when we had a parentActivity set in the manifest file. We could add one back there that references the BrowserActivity and it would likely work after manifest merging, but that feels off to me. Let me know what you think 👍