fix: allow enabling/disabling of FireActivity transition animation
The FireActivity was always launching with a fade in/out transition, this meant the user would see the previous tab briefly.

We can add an option to enable/disable the transition and then call it without specifically when the fire button is pressed. The other areas where we clear data remain unchanged in terms of the transition.