fix: visibility of navigation bar buttons
I noticed that on lower API levels (26,27) tha the navigation bar was black and the buttons were not visible.

This change makes the navigation bar transparent and the dialog edge to edge, retaining the current nav bar colour and letting the system decide on how to overlay the navigation bar.