Tab manager scrolling fix (#4918)
Task/Issue URL:
https://app.asana.com/0/488551667048375/1208062551083727/f

### Description

This PR fixes the initial scroll position after a tab manager is opened.
The position should now be centered to the current tab.

The problem was that the call to scroll to the selected tab was in the
`render()` function, which would not always happen before the
`LayoutManager` was initialized because the `LiveData` observers were
called in different/inconsistent order. This was causing the tabs to
sometimes show the first tab when initially opening the tab manager.

The PR also improves the scroll position when switching between the list
mode and grid mode, which should now be pretty much centered around the
same middle item.

### Steps to test this PR

I was able to reproduce this only when I had a lot of tabs open. If you
don’t have many, you can add this to the `TabSwitcherViewModel`:

```kotlin
    init {
        viewModelScope.launch(dispatcherProvider.io()) {
            if (tabRepository.flowTabs.first().size < 100) {
                repeat(100) {
                    tabRepository.add("cnn.com")
                }
            }
        }
    }
``` 

_Initial position_
- [x] Go to tab manager
- [x] Select a tab somewhere in the middle
- [x] Leave the tab manager
- [x] Go back
- [x] Notice the scroll position is centered on the selected tab

_Configuration change_
- [x] Go to tab manager
- [x] Select the first tab
- [x] Go back to the tab manager
- [x] Scroll away from the selected tab
- [x] Rotate the device
- [x] Notice the scroll position has not changed

_Switching view mode_
- [x] Go to tab manager
- [x] Scroll somewhere where you can identify the item in the middle
- [x] Change the view mode
- [x] Notice the previous item is still roughly in the middle
- [x] Switch the view mode again
- [x] Notice the position is roughly the same again