fix: always set status and navigation icon colour
When the animation plays we have a black background. In light mode this leads to the status bar and navigation bar icon colours being too dark.

Instead we can ensure they are always set to light so that the icons and buttons are always visible.

This contrast works well across all animations.

As in dark mode they were already white, they are no changes there.