VersionCatalog with version-catalog-update plugin
I wanted to see if there were alternative plugins that might let us retain more control in our VersionCatalog.

I found this buried in my github stars. This plugin builds on top of the (classic) Ben Manes "Versions" plugin to run on a VersionCatalog: https://github.com/littlerobots/version-catalog-update-plugin

The version-catalog-update plugin:
  - Strips out anything without a version, which is why you see "# @keep" comments for kotlin and java plugins
  - Wants to use it's own layout when updating. Notice with moved to one long list rather than keeping our own spacing.
  - Would need an extra configuration block to ensure we stay on stable version updated but totally doable. I left it out for demo purposes.
  - Automatically added a unified version for com.frybits.harmony. Maybe we should be doing that but currently they are on different versions.
  - You can use comments to pin versions if you want e.g. we could pin RxJava versions

If I had to pick one of the two I'd pick this plugin as it keeps our own ordering at least and does not change the dependency naming. But honestly I don't like the fact that both plugins are changing the file and not just the versioning. The Github apps, Renovate (and maybe Dependabot) do not make any changes to the file other than the version number.