always observe specificPageUrl
Due to our agreed behaviour change we now always need to have the specificPageUrl displayed to the user so we always need to get it from the store

We can remove the showKeyboard() function as well so that it only shows if the user taps the input

We now set the url when in onStop otherwise we'll need to deal with debouncing due to continuously updating state. This is simpler.

We need to use the app scope as the viewModelScope will be cancelled