WIP
Some terrible terrible JC to hack this together. With this commit it should

- Play shield animation on launch (proving Lottie works even though we may need a new animation as i had to hack the scale up to get it to fit)
- Animates expansion and contraction of the url bar when activated/deactivated. This is not perfect but looks decent enough for a "free" animation with animateContentSize()
- Adds the blue border when activated
- Selects all text when activated for easier deletion
- Has click handling for the fire button and tabs. The fire animation loads the old web page so that would need fixing but the tabs get wiped so I'm happy about that.
- I added an increment and decrement animation for the tab icon
- I avoided the overflow menu but not because I thought it would be difficult, just did not have enough time to move it over

If I had to do this again I would really consider how to do the TextField. Ideally we'd used the Material3 version for this, but they have some annoying requirements around icon sizes (I got around the padding of the text issue more easily) This puts the TextField at 48dp, ours is 40dp. Now we could just copy paste the Material implementation and rip out what we don't need. That comes with maintenance but gives us a lot of the stuff that I've hacked together for free (e.g. all the managing of state) Or we could write it properly from "scratch", basically better than what I did here but with manual state management. It's a tough call. I think I'd be tempted to copy the Material implementation to try again in future as a base, and see how far we could get with it.

This, unfortunately, is one of the downsides of JC, Google can be rigid when it comes to it's Material Composables, much like the View counterparts and it's not unheard of for people to copy implementations and edit them to get the behaviour they want. And in fact I've seen comments from Googlers that don't discourage this=, I think with JC they expect more people to "get their hands dirty" so to speak and not to be afraid of either copying and adjusting or even writing your own custom Composable.

Let's be honest, the omnibar is one of our more complex UI components, so this was going to be a tough ask, but hopefully this is good enough to prove we could leave some of the BrowserFragment alone while still moving over parts to JC if we wanted. I don't fear moving Settings for example.