open show on app launch setting when set
The launchNewSearchOrQuery seemed to be the best place to add check if we should launch the show on app setting. It does a lot of checking based on what's passed into the intent as far as I can see.

If there is no specialised launch happening then we check the show on app launch setting

1. If LastOpenedTab is set, we do nothing
2. If NewTabPage is set then we open a new tab
3. If SpecificPage is set we compare the url to the current selected tab, if it's not the same we open a new tab otherwise we do nothing

I had to expose selectedTab from the TabsDao so I could the tab in a synchronous fashion to check. I would have made the tabsDao.selectedTab() suspending but that would have meant changes elsewhere.