Add Feature Inventory screen (internal) (#5001)
Task/Issue URL:
https://app.asana.com/0/1198194956794324/1207827226270692/f

### Description
Adds a Feature Flag Inventory screen in internal settings.

Incidentally, the PR had to fix some tests and took the opportunity to
use the `FakeFeatureToggleFactory` rather than mocking the feature
flags.
Because of that I also had to add update/add some lint-baselines.xml

### Steps to test this PR
- [ ] install from this branch and launch the app
- [ ] go to Settings -> scroll down -> Feature Flag Inventory
- [ ] verify the list of feature flags appear
- [ ] use the search box to search for features
- [ ] verify filtering is done correctly
- [ ] toggle one of the features ON and/or OFF
- [ ] verify the toggles actually toggle the feature state and whatever
code guarded by the feature reacts properly
- [ ] tap on the feature name
- [ ] verify the `Feature Flag Info` bottom sheet pops up and contains
the FF information (this info is only present for features that are in
the remote config)