Fix survey's store declaration (#5014)
Task/Issue URL:
https://app.asana.com/0/608920331025315/1208317909207871/f

### Description
To support saving the feature's settings for autofill surveys, we added
a `AutofillSurveyFeatureSettingsStore`. However, the config of the
autofill feature wrongly stated which settings store to use:
- specified: `AutofillSiteBreakageReportingRemoteSettingsPersister`
- should have been: `AutofillSurveyFeatureSettingsStore`

This doesn't matter in terms of functionality. i.e., it works as
expected (hence why it wasn't caught in testing), but still good to tidy
up.

### Steps to test this PR
- QA optional