Fix text glitches to new line in TypeAnimationTextView (#4874)
Task: https://app.asana.com/0/1125189844152671/1208000645162922/f

Issue URL: #4869

### Description

There is a typing animation that adds each letter to the screen
sequentially. The last word of each line can jerk to the next line if
the word is long enough. On iOS the equivalent animation keeps
characters in position as they are revealed. On Android this animation
is more akin to what you would see when you are actually typing and you
type a long word at the end of a line. The word jerks to the next line.
This is fine when you are typing. But at speed an in an animation it
does not look good. This PR copies iOS style, and applies 100%
transparency to all letters that have not yet been "typed" in the
animation.

### Steps to test this PR

1. Install a fresh copy of the app
2. Do onboarding

_Feature 1_
- [ Change TypeAnimationTextView to progressively reveal each character
rather than progressively adding each character to the output sting ]

### UI changes

Before


[Screen_recording_20240807_211447.webm](https://github.com/user-attachments/assets/8e619a1f-5cc1-46ac-8080-e0d940aa3ed4)

After


[Screen_recording_20240808_021820.webm](https://github.com/user-attachments/assets/f49489f0-404e-41eb-af94-09f0bade0c9a)

---------

Co-authored-by: bemusementpark <bemusementpark>