Prevent system autofill offering to save passwords (#5091)
Task/Issue URL:
https://app.asana.com/0/1203822806345703/1208458715174512/f

### Description
On Android 14, `importantForAutofill` is not respected so need to do
something else to stop the system autofill provider prompt offering to
save a password that our user has just manually entered or edited

### Steps to test this PR
QA-optional. But if you are curious:

- [ ] Install on Android 14 device, with Google Password Manager set as
autofill provider
- [ ] Manually add a password in our password management screen and save
it. Return to the password management list
- [ ] Verify you aren't prompted to save in Google Password Manager