resolve full url if possible
The new ShowOnAppLaunchUrlFetcher uses our api OkHttpClient to make a call to get the full resolved URL. If anything goes wrong, we just fall back to the url that the user entered and never retry. We're operating on the basis that a user likely has an internet connection.

We need to ensure that the scheme is http or https otherwise we might get a failure when making the call (e.g. ftp)