update selectTabByUrl behaviour
the query will now match partial urls and not just the fully resolved ones.

e.g. duckduckgo.com will match with https://www.duckduckgo.com/

selectTabByUrl is used elsewhere for favourites in widgets and shortcuts where we'll be using the fully resolved url anyway

we now order descending by position when picking the tab, we may have multiple so it makes sense that we pick the last tab we find which is likely to have been a more recent tab for the user