update ShowOnAppLaunchOptionHandler to check a user and resolved url
Now we're storing both the url and resolved urls, we need to check for duplicate tabs.

This also required a slight tweak in the behaviour in how we check the urls in the db. We have a new query that allows us to mathc on urls by not caring about the scheme and subdomain, and only focusing on what comes after. The name isn't great so open to suggestions! I would have liked to have done this in a more type safe way rather than a String.