handle storing of resolved urls
We add a function to our Handler (should it be manager? 🤷) and if we know that we're on the root of the tab, we have a matching tab id with what was stored when we launched the tab, and it's not null, then we can set the resolved url in the store and re-use it for future launches to avoid duplicate tabs

Next up, tabs!