Omnibar: Url alignment not focused mode (#5149)
Task/Issue URL:
https://app.asana.com/0/1174433894299346/1208542811348566

### Description
When the omnibar is not focused we want to show the beginning of the
url, not the end of it.

### Steps to test this PR

Note: This needs to be tested on both `LegacyOmnibarView` and
`OmnibarLayout`. To change between the two open the Feature Flag
Inventory and flip the refactor flag on / off (on = `OmnibarLayout` /
off = `LegacyOmnibarView`)

- [x] Navigate to any site with a long enough URL
- [x] Verify that when the omnibar is focused, the text is highlighted
and you can see the end of it
- [x] Take focus away from the omnibar
- [x] Verify that the text is not highlighed, and you can see the start
of the url

### UI changes
| Before  | After |
| ------ | ----- |

![Screenshot_20241017_164401](https://github.com/user-attachments/assets/8a343fd0-9985-4a2c-9049-b0100ff38014)|![Screenshot_20241017_164330](https://github.com/user-attachments/assets/7df12f64-08c6-4233-9e4e-d81c6375d53c)|