add and fix GeneralSettingsViewModelTest
I noticed that the initial ViewState in GeneralSettingsViewModel was never actually emitting, because we relied on suspending functions to determine some of the state. We have to mock these functions in setup so that we actually get the first ViewState emission.

As we need greater control over when GeneralSettingsViewModel is initialised for some of the tests I extracted it to a function that we can call when needed rather than in setup.